requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==5.0.9
Flask==1.1.2
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
motor==2.3.1
numpy==1.20.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.15.1
ptyprocess==0.7.0
Pygments==2.9.0
PyJWT==2.0.1
pymongo==3.11.3
pymsteams==0.1.14
pyparsing==2.4.7
python-dateutil==2.8.1
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
smopy==0.0.7
staticmap==0.5.5
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
